pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "pyagent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40",
    "click>=8.1",
    "docstring_parser>=0.16",
    "google-genai>=0.8",
    "openai>=1.50",
    "petname>=2.6",
    "platformdirs>=4",
    "requests>=2.31",
    "rich>=13",
]

[project.scripts]
pyagent = "pyagent.cli:main"
pyagent-skills = "pyagent.skills_cli:main"
pyagent-sessions = "pyagent.sessions_cli:main"

[tool.setuptools.packages.find]
include = ["pyagent*"]

[tool.setuptools.package-data]
pyagent = [
    "defaults/*.md",
    "skills/*/SKILL.md",
    "skills/*/*.py",
]
